Move NullVaultExtension to SPI
